+++ parser_library/src/diagnostic.cpp
@@ -2708,10 +2708,22 @@ diagnostic_s diagnostic_s::warning_L0006(const utils::resource::resource_locatio
         diagnostic_tag::none);
 }
 
+diagnostic_s diagnostic_s::fade(const utils::resource::resource_location& loc, const range& range)
+{
+    return diagnostic_s(loc.get_uri(),
+        range,
+        diagnostic_severity::hint,
+        "PREPROC",
+        "Statement overwritten by preprocessor",
+        {},
+        diagnostic_tag::unnecessary); // todo move and make more verbose
+}
+
 diagnostic_op diagnostic_op::error_S100(std::string_view message, const range& range)
 {
     return diagnostic_op(diagnostic_severity::error, "S100", concat("Long ordinary symbol name - ", message), range);
 }
+
 std::string diagnostic_decorate_message(std::string_view field, std::string_view message)
 {
     static const std::string_view prefix = "While evaluating the result of substitution '";