+++ parser_library/src/processing/instruction_sets/asm_processor.h
@@ -39,10 +39,18 @@ public:
 
     void process(std::shared_ptr<const processing::resolved_statement> stmt) override;
 
+    static bool parse_copy(analyzing_context ctx,
+        workspaces::parse_lib_provider& lib_provider,
+        context::id_index copy_member_id,
+        const range& operand_range,
+        const range& stmt_range,
+        diagnosable_ctx* diagnoser);
+
     static bool process_copy(const semantics::complete_statement& stmt,
         analyzing_context ctx,
         workspaces::parse_lib_provider& lib_provider,
-        diagnosable_ctx* diagnoser);
+        diagnosable_ctx* diagnoser,
+        bool enter_copy = true);
 
 private:
     opencode_provider* open_code_;