+++ parser_library/src/processing/processing_manager.cpp
@@ -42,6 +42,7 @@ processing_manager::processing_manager(std::unique_ptr<opencode_provider> base_p
     , opencode_prov_(*base_provider)
     , lsp_analyzer_(*ctx_.hlasm_ctx, *ctx_.lsp_ctx, file_text)
     , stms_analyzers_({ &lsp_analyzer_ })
+    , file_loc_(file_loc)
 {
     switch (data.proc_kind)
     {
@@ -249,7 +250,24 @@ void processing_manager::finish_copy_member(copy_processing_result result)
     lsp_analyzer_.copydef_finished(member, std::move(result));
 }
 
-void processing_manager::finish_opencode() { lsp_analyzer_.opencode_finished(); }
+void processing_manager::finish_opencode()
+{
+    for (const auto& stmt : opencode_prov_.get_preprocessor_statements())
+    {
+        if (!stmt)
+            continue;
+
+        lsp_analyzer_.analyze(*stmt);
+
+        if (stmt->m_resemblence == context::id_storage::well_known::COPY)
+            asm_processor::process_copy(*stmt, ctx_, lib_provider_, nullptr, false);
+
+        // diagnosable_impl::add_diagnostic(diagnostic_s::fade(file_loc_, stmt->stmt_range_ref())); // todo create
+        // separate 'fade' container
+    }
+
+    lsp_analyzer_.opencode_finished();
+}
 
 void processing_manager::start_macro_definition(
     macrodef_start_data start, std::optional<utils::resource::resource_location> file_loc)